# swagger/v1/swagger.yaml
openapi: 3.0.1
info:
  title: API V1
  version: v1
paths:
  /api/v1/locations:
    get:
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - in: query
          name: per_page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      summary: Retrieves all locations
      responses:
        '200':
          description: A list of locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Location'
  /api/v1/locations/{id}:
    get:
      summary: Retrieves a single location by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          description: ID of the location to retrieve
      responses:
        '200':
          description: The location
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '404':
          description: Location not found
  /api/v1/locations/{id}/air_qualities:
    get:
      summary: Retrieves air qualities for a location
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          description: ID of the location to retrieve air qualities for
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - in: query
          name: per_page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Number of items per page for pagination
      responses:
        '200':
          description: Air qualities for the location
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AirQuality'
        '404':
          description: Location not found
components:
  schemas:
    Location:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: The unique identifier of the location
        name:
          type: string
          description: The name of the location
        latitude:
          type: number
          format: float
          description: The latitude coordinate of the location
        longitude:
          type: number
          format: float
          description: The longitude coordinate of the location
      required:
        - id
        - name
        - latitude
        - longitude

    AirQuality:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: The unique identifier of the air quality data
        location_id:
          type: integer
          format: int64
          description: The ID of the location associated with the air quality data
        pollutant:
          type: string
          description: The type of pollutant measured (e.g., CO2, PM2.5)
        value:
          type: number
          format: float
          description: The value of the pollutant measurement
        timestamp:
          type: string
          format: date-time
          description: The timestamp when the measurement was recorded
      required:
        - id
        - location_id
        - pollutant
        - value
        - timestamp